requirements.txt
streamlit==1.31.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
google-generativeai==0.3.2
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
openai-whisper==20231117
assemblyai==0.20.0
textblob==0.17.1
vaderSentiment==3.3.2
sentence-transformers==2.3.1
mediapipe==0.10.14
opencv-python==4.9.0.80
python-dotenv==1.0.0
pandas==2.2.0
numpy==1.26.3
pillow==10.2.0
streamlit-authenticator==0.2.3